requirements.txt
aiogram~=3.3.0
bs4~=0.0.1
beautifulsoup4~=4.12.2
selenium~=4.16.0
python-dotenv~=1.0.0
fake-useragent~=1.4.0